Refactor scipp.signal.butter to take coord as argument directly
